+++ trunk/vxquery/vxquery-core/src/main/java/org/apache/vxquery/runtime/functions/node/FnDocScalarEvaluatorFactory.java
@@ -58,7 +58,7 @@ public class FnDocScalarEvaluatorFactory extends AbstractTaggedValueArgumentScal
         final DataInputStream di = new DataInputStream(bbis);
         final int partition = ctx.getTaskAttemptId().getTaskId().getPartition();
         // TODO Add the data source scan id.
-        final ITreeNodeIdProvider nodeIdProvider = new TreeNodeIdProvider(partition);
+        final ITreeNodeIdProvider nodeIdProvider = new TreeNodeIdProvider((short) partition);
 
         return new AbstractTaggedValueArgumentScalarEvaluator(args) {
             @Override
